Technical Test Finished
Features requested to implement:

-The app needs to be able to add and remove patients.
-When a patient is selected a new screen should be presented displaying
the fetal monitor (graph of the babies heart rate on top and the mothers
on the bottom)
-The graph needs simulate real time and as such should present a
simulated heart rate over time at resolution of 10 samples per second
where the most recent sample is at the right of the screen and the
oldest is at the left of the screen.
-The two graphs (the babies and the mothers) should animate/move
together.
-The user of the app should be able to scroll the monitor back in time.